First stab at implementing Vitter’s Method A
When the size of the cartesian product is large, we’ll want to
sample it (without replacement). My idea is to implement Method D
from [0], but to do that, I need to implement the simpler Method A
first.

This commit contains what I think is a roughly clojurified translation
of the pseudocode for Method A from Vitter’s paper (assuming the
sampled set being [0..max) and not special-casing the last step
just yet).

I think it’s buggy. Running (method-a 100 200) should return 100
numbers evenly distributed between 0 and 200, so the max of such
sequence should be close to 200; but in practice I’m getting a
max of something like 161, which seems very unlikely.

Despite this, I’m committing this half-baked approach. Commit early,
commit often, document learnings in the commit description: that’s
how I want to roll in this library. Next steps: lay my hands on
an actual implementation, dig into it and see where it diverges
from mine.

 [0]: Vitter, Jeffrey Scott. “An Efficient Algorithm for Sequential
      Random Sampling”. ACM transactions on mathematical software
      (TOMS) 13.1 (1987): 58-67.
      https://www.ittc.ku.edu/~jsv/Papers/Vit87.RandomSampling.pdf