ClojureScriptize the code
This was painful.

I had to provide separate CLJS implementations of CartesianProduct and
CartesianProductSeq (cljs protocols and clj Java interfaces are
mutually incompatible), so there’s some code duplication. (I didn’t
bother implementing IEquiv for the cljs version.)

Getting the with-combination macro to work cross-platform was
tricky: requiring any external namespaces from the macro ns
(cartesian.test) is a no-no, the macro must use fully-qualified
namespaces, and it must resort to a dirty trick to detect
if we’re in the Clojure or ClojureScript realm.